refactor(arch): finalize project-wide modularization
  details:
   - decomposed ItemFeatureService and PipelineExecutor into focused sub-modules
   - resolved ResilienceConfig naming collisions by introducing ResilienceMetricsConfig
   - unified CircuitBreakerConfig definitions into a single source of truth
   - corrected AnalyticsSidecar initialization to use configured ClickHouse clients
   - updated benchmarks to support new PipelineRegistry and configuration structures
   - enforced pure manifest pattern across every mod.rs file in the workspace
   - eliminated all compiler warnings and type mismatches project-wide

  impact:
   - ensures 100% architectural obedience to modularity and SRP standards
   - eliminates cognitive load by bringing all implementation files under 300 lines
   - restores high-fidelity performance measuring through stabilized benchmarks
   - provides a clean, production-ready foundation for the Bongas-AI engine

+++ src/circuit_breaker/breaker/handlers.rs
@@ -155,7 +155,7 @@ impl CircuitBreaker {
     pub(super) fn maybe_trip(&self, consecutive_failures: u32) {
         // Check consecutive failure threshold first (faster)
         if let Some(threshold) = self.config.consecutive_failure_threshold() {
-            if consecutive_failures >= threshold {
+            if u64::from(consecutive_failures) >= threshold {
                 self.trip_circuit();
                 return;
             }
